+++ amazon/core/src/main/kotlin/org/http4k/connect/amazon/model/messageFields.kt
@@ -1,0 +1,13 @@
+package org.http4k.connect.amazon.model
+
+interface MessageFields {
+    fun toFields(index: Int): Map<String, String>
+}
+
+enum class DataType {
+    String, Number, Binary
+}
+
+fun asList(vararg messageFields: List<MessageFields>) = messageFields.flatMap {
+    it.flatMapIndexed { index, messageFields -> messageFields.toFields(index + 1).toList() }
+}