+++ amazon/s3/client/src/main/kotlin/org/http4k/connect/amazon/s3/action/ListBuckets.kt
@@ -6,6 +6,8 @@ import org.http4k.connect.Http4kConnectAction
 import org.http4k.connect.Listing
 import org.http4k.connect.RemoteFailure
 import org.http4k.connect.amazon.model.BucketName
+import org.http4k.connect.amazon.model.text
+import org.http4k.connect.amazon.model.xmlDoc
 import org.http4k.core.Method
 import org.http4k.core.Request
 import org.http4k.core.Response
@@ -19,8 +21,8 @@ class ListBuckets : S3Action<Listing<BucketName>> {
     override fun toResult(response: Response) = with(response) {
         when {
             status.successful -> {
-                val buckets = documentBuilderFactory().parse(body.stream).getElementsByTagName("Name")
-                val items = (0 until buckets.length).map { BucketName.of(buckets.item(it).textContent) }
+                val buckets = xmlDoc().getElementsByTagName("Name")
+                val items = (0 until buckets.length).map { BucketName.of(buckets.item(it).text()) }
                 Success(if (items.isNotEmpty()) Listing.Unpaged(items) else Listing.Empty)
             }
             else -> Failure(RemoteFailure(Method.GET, Uri.of("/"), status))