+++ amazon/s3/client/src/main/kotlin/org/http4k/connect/amazon/s3/action/S3Action.kt
@@ -1,13 +1,5 @@
 package org.http4k.connect.amazon.s3.action
 
 import org.http4k.connect.Action
-import org.xml.sax.InputSource
-import java.io.StringReader
-import javax.xml.parsers.DocumentBuilderFactory
 
 interface S3Action<R> : Action<R>
-
-internal fun documentBuilderFactory() =
-    DocumentBuilderFactory.newInstance()
-        .newDocumentBuilder()
-        .apply { setEntityResolver { _, _ -> InputSource(StringReader("")) } }