+++ amazon/sns/client/src/main/kotlin/org/http4k/connect/amazon/model/sns.kt
@@ -1,0 +1,17 @@
+package org.http4k.connect.amazon.model
+
+import dev.forkhandles.values.StringValue
+import dev.forkhandles.values.StringValueFactory
+import dev.forkhandles.values.minLength
+
+class SNSMessageId private constructor(value: String) : StringValue(value) {
+    companion object : StringValueFactory<SNSMessageId>(::SNSMessageId, 1.minLength)
+}
+
+class PhoneNumber private constructor(value: String) : StringValue(value) {
+    companion object : StringValueFactory<PhoneNumber>(::PhoneNumber, 1.minLength)
+}
+
+class TopicName private constructor(value: String) : ResourceId(value) {
+    companion object : StringValueFactory<TopicName>(::TopicName, 1.minLength)
+}