+++ amazon/sns/fake/src/main/kotlin/org/http4k/connect/amazon/sns/model.kt
@@ -1,0 +1,11 @@
+package org.http4k.connect.amazon.sns
+
+import org.http4k.connect.amazon.model.ARN
+import org.http4k.connect.amazon.model.SNSMessageId
+import org.http4k.template.ViewModel
+
+object DeleteTopicResponse : ViewModel
+data class CreateTopicResponse(val topicArn: ARN) : ViewModel
+data class ListTopicsResponse(val arns: List<ARN>) : ViewModel
+data class PublishResponse(val messageId: SNSMessageId) : ViewModel
+