+++ amazon/sns/fake/src/test/kotlin/org/http4k/connect/amazon/sns/FakeSNSTest.kt
@@ -1,0 +1,7 @@
+package org.http4k.connect.amazon.sns
+
+import org.http4k.connect.amazon.fakeAwsEnvironment
+
+class FakeSNSTest : SNSContract(FakeSNS()) {
+    override val aws = fakeAwsEnvironment
+}