+++ amazon/sqs/client/src/main/kotlin/org/http4k/connect/amazon/sqs/action/CreateQueue.kt
@@ -5,6 +5,9 @@ import dev.forkhandles.result4k.Success
 import org.http4k.connect.Http4kConnectAction
 import org.http4k.connect.RemoteFailure
 import org.http4k.connect.amazon.model.QueueName
+import org.http4k.connect.amazon.model.Tag
+import org.http4k.connect.amazon.model.text
+import org.http4k.connect.amazon.model.xmlDoc
 import org.http4k.core.Method.POST
 import org.http4k.core.Response
 import org.http4k.core.Uri
@@ -14,13 +17,13 @@ import java.time.format.DateTimeFormatter.ISO_ZONED_DATE_TIME
 @Http4kConnectAction
 data class CreateQueue(
     val queueName: QueueName,
-    val tags: Map<String, String> = emptyMap(),
+    val tags: List<Tag> = emptyList(),
     val attributes: Map<String, String> = emptyMap(),
     val expires: ZonedDateTime? = null
 ) : SQSAction<CreatedQueue>(
     "CreateQueue",
-    *(tags.entries
-        .flatMap { listOf("Tag.Key" to it.key, "Tag.Value" to it.value) } +
+    *(tags
+        .flatMap { listOf("Tag.Key" to it.Key, "Tag.Value" to it.Value) } +
         attributes.entries
             .flatMapIndexed { i, it ->
                 listOf(
@@ -49,8 +52,6 @@ data class CreatedQueue(
 ) {
     companion object {
         fun from(response: Response) =
-            with(documentBuilderFactory().parse(response.body.stream)) {
-                CreatedQueue(Uri.of(text("QueueUrl")))
-            }
+            CreatedQueue(Uri.of(response.xmlDoc().text("QueueUrl")))
     }
 }