+++ amazon/sqs/client/src/main/kotlin/org/http4k/connect/amazon/sqs/action/SQSAction.kt
@@ -8,12 +8,6 @@ import org.http4k.core.Uri
 import org.http4k.core.body.form
 import org.http4k.core.with
 import org.http4k.lens.Header.CONTENT_TYPE
-import org.w3c.dom.Document
-import org.w3c.dom.Node
-import org.w3c.dom.NodeList
-import org.xml.sax.InputSource
-import java.io.StringReader
-import javax.xml.parsers.DocumentBuilderFactory
 
 abstract class SQSAction<R>(private val action: String, private vararg val mappings: Pair<String, String>?) :
     Action<R> {
@@ -29,21 +23,3 @@ abstract class SQSAction<R>(private val action: String, private vararg val mappi
 
     protected abstract fun uri(): Uri
 }
-
-internal fun documentBuilderFactory() =
-    DocumentBuilderFactory.newInstance()
-        .newDocumentBuilder()
-        .apply { setEntityResolver { _, _ -> InputSource(StringReader("")) } }
-
-internal fun Document.text(name: String) = textOptional(name)!!
-
-internal fun Document.textOptional(name: String) = getElementsByTagName(name).item(0)?.textContent?.trim()
-
-internal fun NodeList.sequenceOfNodes(onlyChildrenNamed: String? = null): Sequence<Node> {
-    var i = 0
-    val baseSequence = generateSequence { i.let { if (it == length) null else item(it) }.also { i++ } }
-    return onlyChildrenNamed?.let { baseSequence.filter { it.nodeName == onlyChildrenNamed } } ?: baseSequence
-}
-
-internal fun Node.firstChild(name: String) = children(name).first()
-internal fun Node.children(name: String) = childNodes.sequenceOfNodes(name)