+++ amazon/sqs/client/src/main/kotlin/org/http4k/connect/amazon/sqs/action/SendMessage.kt
@@ -8,6 +8,10 @@ import org.http4k.connect.amazon.model.ARN
 import org.http4k.connect.amazon.model.AwsAccount
 import org.http4k.connect.amazon.model.QueueName
 import org.http4k.connect.amazon.model.SQSMessageId
+import org.http4k.connect.amazon.model.asList
+import org.http4k.connect.amazon.model.text
+import org.http4k.connect.amazon.model.textOptional
+import org.http4k.connect.amazon.model.xmlDoc
 import org.http4k.core.Method
 import org.http4k.core.Response
 import org.http4k.core.Uri
@@ -77,7 +81,7 @@ data class SentMessage(
 ) {
     companion object {
         fun from(response: Response) =
-            with(documentBuilderFactory().parse(response.body.stream)) {
+            with(response.xmlDoc()) {
                 SentMessage(
                     text("MD5OfMessageBody"),
                     SQSMessageId.of(text("MessageId")),
@@ -85,8 +89,4 @@ data class SentMessage(
                 )
             }
     }
-}
-
-private fun asList(vararg messageFields: List<MessageFields>) = messageFields.flatMap {
-    it.flatMapIndexed { index, messageFields -> messageFields.toFields(index + 1).toList() }
 }